+++ CHANGELOG.md
@@ -106,6 +106,7 @@ More details about these changes can be found on our GitHub repo.
 * dns_rfc2136 plugin now supports explicitly specifing an authorative
   base domain for cases when the automatic method does not work (e.g.
   Split horizon DNS)
+* NGINX plugin now supports using HTTP2 when rewriting configuration.
 
 ### Changed